// src/app/core/models/website-setting.model.ts
export interface WebsiteSetting {
  websiteSettingId: number;
  companyName: string;
  logoUrl: string | null;
  faviconUrl: string | null;
  email: string | null;
  phone: string | null;
  whatsAppNumber: string | null;
  address: string | null;
  facebookUrl: string | null;
  instagramUrl: string | null;
  twitterUrl: string | null;
  linkedInUrl: string | null;
  youTubeUrl: string | null;
  footerText: string | null;
  updatedDate: string | null;
}